requirements.txt
# Web Framework
flask>=2.0.0
gunicorn>=20.0.0

# HTTP & Web Scraping
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
selenium>=4.0.0

# Data Processing
pandas>=1.5.0
numpy>=1.23.0

# NLP & Text Processing
jieba>=0.42.1
scikit-learn>=1.1.0

# Database Drivers
psycopg2-binary>=2.9.0
pymongo>=4.3.0
redis>=4.0.0
elasticsearch>=7.17.0
neo4j>=5.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Data Validation
pydantic>=2.0.0

# SSH/Remote
paramiko>=2.11.0

# Utilities
python-dateutil>=2.8.0

# Task Scheduling
apscheduler>=3.10.0

# Rate Limiting
flask-limiter>=3.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
